LPD-80539 SF, move package
