+++ modules/apps/feature-flag/feature-flag-web/src/main/java/com/liferay/feature/flag/web/internal/model/listener/CompanyModelListener.java
@@ -5,7 +5,7 @@
 
 package com.liferay.feature.flag.web.internal.model.listener;
 
-import com.liferay.feature.flag.web.internal.DeprecationFeatureFlagUtil;
+import com.liferay.feature.flag.web.internal.util.DeprecationFeatureFlagUtil;
 import com.liferay.feature.flag.web.internal.feature.flag.FeatureFlagsBagProvider;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
